Setup Tox and Travis Build Matrix
Summary:
The way I've done build matrices in the past is to use Tox to handle all
the tedious work of setting up virtul aenvironments and running the
tests.

It also looks like there's a nice plugin,
[tox-travis](https://github.com/tox-dev/tox-travis), which takes in the
evironment definsed to run the appropriate tests.

Unfortunately, when this was run against Pythonn 3.5, the tests failed,
so that's a later PR.

Test Plan:
- Run `tox` locally and confirm the tests pass on py36, py37, and py38