Support python3.10
Similar to #48 (and #47), update `requirements.txt` with
compatible `werkzeug` dependency. Update to a more recent Flask version
and fix a compatibility error with the new jinja2 transitive dependency.

Also pin `jinja2` to major version, just in case.

And finally, add a small compatibility patch to support python3.10.

Tested on python3.10 and python3.9 virtualenvs.