Add actions run on PR, and mac
Previously actions was only running on pushes. Tidy that up and make it
run on PR's as well.

Run tests on mac as well- added a windows run to the matrix, but it
fails due to path mismatch issues, which are addressed in other PRs, so
leaving it disabled here.